+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Changed
 - Documentation
+- Code comments
 
 ## [0.3.0] - 2021-08-08
 ### Added