+++ CHANGELOG.md
@@ -5,7 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-### Added
+
+## [0.4.0] - 2021-08-11
+### Added 
 - CLI option for **process** operation to use a folder as Markdown input. [#4]
 - Dark Theme [#2]